Targeted update - refactoring
